requirements.txt
python-dotenv
streamlit
langchain
langchain-community
langchain-groq
langchain-huggingface
faiss-cpu
sentence-transformers
beautifulsoup4